pyproject.toml
[project]
name = "shark-species-classifier"
version = "0.1.0"
description = "Shark species image classification (13 classes) with Hydra + Lightning + MLflow + DVC"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Timur Burov" }]
keywords = ["computer-vision", "classification", "mlops", "hydra", "pytorch-lightning", "mlflow", "dvc"]

dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "pytorch-lightning>=2.2",
  "torchmetrics>=1.3",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "numpy>=1.26",
  "pillow>=10.1",
  "scikit-learn>=1.3",
  "tqdm>=4.66",
  "mlflow>=2.6",
  "dvc>=3.41",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0",
  "ruff>=0.6",
]

[tool.ruff]
line-length = 90
target-version = "py310"
fix = true
show-fixes = true
src = ["shark_species_classifier"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
